requirements.txt
absl-py==2.1.0
addict==2.4.0
asttokens==2.4.1
basicsr==1.4.2
certifi==2022.12.7
chainer==7.8.1
charset-normalizer==2.1.1
decorator==5.1.1
einops==0.8.0
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.13.1
fsspec==2024.2.0
future==1.0.0
grpcio==1.63.0
idna==3.4
imageio==2.34.1
importlib_metadata==7.1.0
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
lazy_loader==0.4
lmdb==1.5.1
lpips==0.1.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyYAML==6.0.1
requests==2.28.1
scikit-image==0.22.0
scipy==1.13.0
sewar==0.4.6
six==1.16.0
stack-data==0.6.3
sympy==1.12
tb-nightly==2.18.0a20240714
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tifffile==2024.5.22
tomli==2.0.1
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.4
traitlets==5.14.3
triton==2.1.0
typing_extensions==4.9.0
urllib3==1.26.13
wcwidth==0.2.13
Werkzeug==3.0.3
yapf==0.40.2
zipp==3.18.1
